Take "hideGlobalTasks" parameter into account (#426)
* Take "hideGlobalTasks" parameter into account

If the hideGlobalTasks parameter is active, the tasks-per-map counter needs to ignore global tasks. We needs to use the same logic that the task list uses to filter out global tasks. The task list checks the map property of the task object for null.

Resolves #424

* Take objectives into account for the tasks-per-map calculation

When all objectes on the current map are already marked as "completed" in the progress store, we can ignore that task for the tasks-per-map counter.

Resolves #424

* Minor code improvement